Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdsa-import-key-sign-p521-expected.txt

Test ECDSA signing operation with a P-521 private key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS signature.byteLength is 132
PASS successfullyParsed is true

TEST COMPLETE
